feat: use max-height 30vh on small screen
On a phone, the image used to be the only thing above the fold. Let's
display an image with close to the original proportions, and also allow
the tagline to appear on the screen below the image.

On the [original template][orig], the image would change to the landscape
variant, and would only occupy the top roughly 25-30% of the screen. The
max-height setting is at 30vh.

[orig]: https://onepagelove.com/demo/split/index.html